refactor : reportservice로 분리 (#981)
