+++ Clubber-External/src/main/java/com/clubber/domain/review/controller/ReviewController.java
@@ -2,6 +2,7 @@ package com.clubber.domain.review.controller;
 
 import com.clubber.domain.club.service.ClubService;
 import com.clubber.domain.domains.review.domain.ReviewSortType;
+import com.clubber.domain.report.service.ReportService;
 import com.clubber.domain.review.dto.*;
 import com.clubber.domain.review.service.ReviewLikeService;
 import com.clubber.domain.review.service.ReviewService;
@@ -22,6 +23,7 @@ public class ReviewController {
 	private final ClubService clubService;
 	private final ReviewService reviewService;
 	private final ReviewLikeService reviewLikeService;
+	private final ReportService reportService;
 
 	@Operation(summary = "동아리 리뷰 동의 여부 반환")
 	@DisableSwaggerSecurity
@@ -79,6 +81,6 @@ public class ReviewController {
 	@Operation(summary = "리뷰 신고")
 	@PostMapping("/report/{id}")
 	public CreateReviewReportResponse postReviewReport(@PathVariable Long id,@RequestBody @Valid CreateReviewReportRequest request){
-		return reviewService.createReviewReport(id,request);
+		return reportService.createReviewReport(id,request);
 	}
 }
